\name{Cambridgecolourdemo}
\alias{Cambridgecolourdemo}
\title{ Function to illustrate the range of Cambridge colours}
\description{
Plots the six University brand colours
}
\usage{
Cambridgecolourdemo()
}
\references{ http://www.cam.ac.uk/cambuniv/style/five.html }
\author{ Andy Lynch }
\seealso{ Camcol }
\keyword{color}
