\name{CRUKcolourdemo}
\alias{CRUKcolourdemo}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Function to show off the CRUK colours}
\description{
plots all thirty branded colours
}
\usage{
CRUKcolourdemo()
}
\references{ www.cancerresearchuk.org/branding/images/pdfs/guidelines }
\author{ Andy Lynch}

\seealso{ \code{\link{CRUKcol}}}
\examples{
CRUKcolourdemo()
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{color}
