\name{redpand-package}
\alias{redpand-package}
\alias{redpand}
\docType{package}
\title{
\packageTitle{redpand}
}
\description{
\packageDescription{redpand}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{redpand}
\packageIndices{redpand}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{redpand}

Maintainer: \packageMaintainer{redpand}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
